More edits to the IMU
I added some more edits to the code. General edits include but are not limited to: Making sure it aligns with the balance board and is in the correct position for each column. Also using the first angle instead of the second, in the case that the second angle is not the correct angle. Also added a debugging tool in the telemetry of the SensorBNO055IMU autonomous. See code for additional changes.